requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
api-connectors==3.1
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.2
binaryornot==0.4.4
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contextlib2==21.6.0
cookiecutter==2.5.0
dataclasses-json==0.5.14
datasets==2.16.1
DateTime==5.4
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
Django==5.0.1
execnet==2.0.2
executing==2.0.1
filelock==3.13.1
flet==0.18.0
flet-core==0.18.0
flet-runtime==0.18.0
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2023.10.0
furl==2.1.0
# Editable install with no version control (gfpgan==1.3.8)
-e i:\desarrollo\proyecto\imagenes_app\gfpgan-master
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==2.113.0
google-api-python-client-py3==1.2

google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0

gptcache==0.1.43
greenlet==3.0.3
gspread==5.12.4
h11==0.14.0
hashring==1.5.1
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.20.3
idna==2.10
importlib-metadata==6.11.0
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.2
bcrypt==4.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.0
langchain-community==0.0.13
langchain-core==0.1.13
langsmith==0.0.83
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
memcache==0.12.0
mock==5.1.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.1
nest-asyncio==1.5.8
numexpr==2.8.8
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
openapi-schema-pydantic==1.2.4
opencv-python==4.9.0.80
orderedmultidict==1.0.1
packaging==23.2
pandas==2.1.4
parso==0.8.3
path==16.10.0
path.py==12.5.0
pdf2image==1.17.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.7
pure-eval==0.2.2
py3-wkhtmltopdf==0.4.1
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydantic_core==2.14.6
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.0.0
PyPDF2==3.0.1
pypng==0.20220715.0
pytest==7.4.4
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
readchar==4.0.5
referencing==0.32.1
repath==0.9.0
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.51
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.30.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torchvision==0.2.2.post3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
ttkthemes==3.2.2
twilio==8.11.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
websocket-client==1.7.0
websockets==11.0.3
wheel==0.42.0
windows-curses==2.3.2
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
